ramnog: Documentation Hub for the 'chef' Ecosystem

ramnog provides vignettes and documentation for the chef ecosystem of R packages for clinical trial evidence generation following AMNOG guidelines. The core functionality is provided by chef, chefStats, and chefCriteria.

Vignettes

Use browseVignettes("ramnog") to see all available vignettes.

See also

Author

Maintainer: Christian Haargaard Olsen cino@novonordisk.com

Authors:

Matthew David Phelps mewp@novonordisk.com

Nicolai Skov Johnsen nosj@novonordisk.com

Henrik Sparre Spiegelhauer hspu@novonordisk.com